% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_urls.R
\name{find_urls}
\alias{find_urls}
\title{Find URLs}
\usage{
find_urls(dates = NULL, s_ad_fct = 1.25)
}
\arguments{
\item{dates}{A date value or vector in the format 'yyyy-mm-dd'
(either of Date or character class), between 04/12/2018 and the present date,
both inclusive. When the date is not found, or there is an
inconsistency in that days' conference, it throws a warning.}

\item{s_ad_fct}{A numeric value to adjust the scope of the dates' search through
the pages in the main website of the conferences.}
}
\value{
A character vector containing the URLs corresponding to the selected
dates.
}
\description{
Finds URLs corresponding to specific dates.
}
\examples{
urls <- find_urls('2020-10-06')
urls <- find_urls(c('2020-10-06','2020-08-19'))
dates <- seq(lubridate::ymd('2020-10-01'),
             lubridate::ymd('2020-10-16'),
             by = 'day')
urls <- find_urls(dates)
}
