% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{text_extraction}
\alias{text_extraction}
\title{Text extraction from URLs}
\usage{
text_extraction(dates = NULL, urls = NULL, start = NULL, end = NULL)
}
\arguments{
\item{dates}{A date value or vector in the format 'yyyy-mm-dd'
(either of Date or character class), between 04/12/2018 and the present date,
both inclusive. When the date is not found, or there is an
inconsistency in that days' conference, it throws a warning. Must be set
to NULL if urls, start or end arguments are specified.}

\item{urls}{A string value or vector containing URLs pointing to conferences hosted in
the website https://presidente.gob.mx/secciones/version-estenografica/.
When the date is not found, or there is an
inconsistency in that days' conference, it throws a warning.
Must be set to NULL if dates, start or end arguments are specified.}

\item{start}{A date value in the format 'yyyy-mm-dd' (either of Date or character class),
between 04/12/2018 and the present date, both inclusive. The function will retrieve
the conferences from that date, inclusive, until the end date, if specified,
or until last available, otherwise. Must be set to NULL if dates or urls
arguments are specified.}

\item{end}{A date value in the format 'yyyy-mm-dd' (either of Date or character class),
between 04/12/2018 and the present date, both inclusive. The function will retrieve
the conferences until that date, inclusive, from the start date, if specified,
or from '2020-12-04', otherwise. Must be set to NULL if dates or urls
arguments are specified.}
}
\value{
A list containing dialog-formatted lists. Every nested list corresponds to
one conference, and every element of it corresponds to an intervention in
that conference (the name of the element corresponds to the one of the interlocutor).
}
\description{
Extracts and formats the stenographic version of the press conferences
for a given date, set of dates, URL, set of URLs,
or for all the conferences available. The name of
the parameter must be specified when calling this function.
}
\examples{
conferences <- text_extraction(dates = c('2020-10-17','2020-10-10'))
}
