% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_parsing.R
\name{urls_parsing}
\alias{urls_parsing}
\title{URLs Parsing}
\usage{
urls_parsing(page)
}
\arguments{
\item{page}{A numeric value to be passed to call
glue::glue('https://presidente.gob.mx/secciones/version-estenografica/page/\{page\}/').}
}
\value{
A character vector containing all the urls stored in the selected page.
}
\description{
The website https://presidente.gob.mx/secciones/version-estenografica/
stores the URLs pointing to the stenographic versions of the presidential
conferences in several pages.
URLs stored in the page n, will be retrieved from the URL
https://presidente.gob.mx/secciones/version-estenografica/page/n/
}
\examples{
# To retrieve all the URLs located in the page five in
# 'https://presidente.gob.mx/secciones/version-estenografica/page/5/',
# call:

urls <- urls_parsing(5)
}
